Reflect changes in cfgrammar and lrtable.

+++ src/lib/parser.rs
@@ -103,7 +103,7 @@ impl<'a, TokId: Clone + Copy + Debug + PartialEq + TryFrom<usize> + TryInto<usiz
       where F: Fn(TIdx) -> u8
     {
         let psr = Parser{rcvry_kind, grm, term_cost: &term_cost, sgraph, stable, lexemes};
-        let mut pstack = vec![StIdx::from(0)];
+        let mut pstack = vec![StIdx::from(0 as u32)];
         let mut tstack: Vec<Node<TokId>> = Vec::new();
         let mut errors: Vec<ParseError<TokId>> = Vec::new();
         let accpt = psr.lr(0, &mut pstack, &mut tstack, &mut errors);
